requirements.txt
beautifulsoup4==4.12.3
datasets==2.20.0
faiss_cpu==1.8.0
fastapi==0.111.0
func_timeout==4.3.5
ipykernel==6.29.4
jieba==0.42.1
json5==0.9.25
jsonlines==4.0.0
jupyter_client==8.6.2
langchain==0.2.5
langchain_community==0.2.5
matplotlib==3.9.0
modelscope_studio==0.4.0
numpy==1.26.4
openai==1.34.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
Pillow==10.3.0
prettytable==3.10.0
pydantic==2.7.4
pytest==7.4.0
python_docx==1.1.2
python_pptx==0.6.23
rank_bm25==0.2.2
Requests==2.32.3
scikit_learn==1.5.0
seaborn==0.13.2
sentence_transformers==3.0.1
setuptools==69.5.1
snowballstemmer==2.2.0
sympy==1.12.1
tiktoken==0.7.0
torch==2.0.0
tqdm==4.66.4
transformers==4.41.2
uvicorn==0.30.1
